requirements.txt
click>=8.1.7
chromadb>=0.5.5
networkx>=3.2
python-docx>=1.1.2
PyMuPDF>=1.24.10
python-pptx>=0.6.23
markdownify>=0.13.1
beautifulsoup4>=4.12.3
pillow>=10.4.0
pytest>=8.3.2
